requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
pyaml==21.10.1
PyYAML==6.0
requests==2.28.0
soupsieve==2.3.2.post1
urllib3==1.26.9
